refactor:PuppyDogAnimationのAnimateを分離
